refactoring to oras pattern
