+++ cmd/notation/cert/generateTest.go
@@ -1,24 +1,75 @@
-package main
+package cert
 
 import (
 	"crypto/rand"
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/pem"
+	"errors"
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/notaryproject/notation-core-go/testhelper"
 	"github.com/notaryproject/notation-go/config"
 	"github.com/notaryproject/notation-go/dir"
+	"github.com/notaryproject/notation/cmd/notation/internal/truststore"
 	"github.com/notaryproject/notation/internal/osutil"
-	"github.com/notaryproject/notation/internal/truststore"
+	"github.com/notaryproject/notation/internal/slices"
 	"github.com/notaryproject/notation/pkg/configutil"
+	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 )
+
+var (
+	keyDefaultFlag = &pflag.Flag{
+		Name:      "default",
+		Shorthand: "d",
+		Usage:     "mark as default",
+	}
+	setKeyDefaultFlag = func(fs *pflag.FlagSet, p *bool) {
+		fs.BoolVarP(p, keyDefaultFlag.Name, keyDefaultFlag.Shorthand, false, keyDefaultFlag.Usage)
+	}
+)
+
+type certGenerateTestOpts struct {
+	name      string
+	bits      int
+	trust     bool
+	isDefault bool
+}
+
+func certGenerateTestCommand(opts *certGenerateTestOpts) *cobra.Command {
+	if opts == nil {
+		opts = &certGenerateTestOpts{}
+	}
+	command := &cobra.Command{
+		Use:   "generate-test [flags] <common_name>",
+		Short: "Generate a test RSA key and a corresponding self-signed certificate.",
+		Args: func(cmd *cobra.Command, args []string) error {
+			if len(args) == 0 {
+				return errors.New("missing certificate common_name")
+			}
+			opts.name = args[0]
+			return nil
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return generateTestCert(opts)
+		},
+	}
+
+	command.Flags().IntVarP(&opts.bits, "bits", "b", 2048, "RSA key bits")
+	command.Flags().BoolVar(&opts.trust, "trust", false, "add the generated certificate to the trust store")
+	setKeyDefaultFlag(command.Flags(), &opts.isDefault)
+	return command
+}
 
 func generateTestCert(opts *certGenerateTestOpts) error {
 	// initialize
 	name := opts.name
+	if strings.TrimSpace(name) == "" {
+		return errors.New("certificate common_name cannot be empty or contain only whitespaces")
+	}
 
 	// generate RSA private key
 	bits := opts.bits
@@ -60,14 +111,14 @@ func generateTestCert(opts *certGenerateTestOpts) error {
 			CertificatePath: certPath,
 		},
 	}
-	err = addKeyCore(signingKeys, keySuite, isDefault)
+	err = addKeyToSigningKeys(signingKeys, keySuite, isDefault)
 	if err != nil {
 		return err
 	}
 	trust := opts.trust
 	// Add to the trust store
 	if trust {
-		if err := truststore.AddCertCore(certPath, "ca", name, true); err != nil {
+		if err := truststore.AddCert(certPath, "ca", name, true); err != nil {
 			return err
 		}
 	}
@@ -104,4 +155,15 @@ func generateCertPEM(rsaCertTuple *testhelper.RSACertTuple) []byte {
 func generateSelfSignedCert(privateKey *rsa.PrivateKey, name string) (testhelper.RSACertTuple, []byte, error) {
 	rsaCertTuple := testhelper.GetRSASelfSignedCertTupleWithPK(privateKey, name)
 	return rsaCertTuple, generateCertPEM(&rsaCertTuple), nil
+}
+
+func addKeyToSigningKeys(signingKeys *config.SigningKeys, key config.KeySuite, markDefault bool) error {
+	if slices.Contains(signingKeys.Keys, key.Name) {
+		return errors.New(key.Name + ": already exists")
+	}
+	signingKeys.Keys = append(signingKeys.Keys, key)
+	if markDefault {
+		signingKeys.Default = key.Name
+	}
+	return nil
 }
