+++ internal/osutil/file.go
@@ -1,7 +1,6 @@
 package osutil
 
 import (
-	"errors"
 	"fmt"
 	"io"
 	"io/fs"
@@ -40,8 +39,8 @@ func WriteFileWithPermission(path string, data []byte, perm fs.FileMode, overwri
 	return file.Close()
 }
 
-// Copy the src file to dst. Existing file will be overwritten.
-func Copy(src, dst string) (int64, error) {
+// CopyToDir copies the src file to dst. Existing file will be overwritten.
+func CopyToDir(src, dst string) (int64, error) {
 	sourceFileStat, err := os.Stat(src)
 	if err != nil {
 		return 0, err
@@ -57,7 +56,7 @@ func Copy(src, dst string) (int64, error) {
 	}
 	defer source.Close()
 
-	if err := os.MkdirAll(dst, 0700); err != nil {
+	if err := os.MkdirAll(dst, 0777); err != nil {
 		return 0, err
 	}
 	certFile := filepath.Join(dst, filepath.Base(src))
@@ -67,64 +66,4 @@ func Copy(src, dst string) (int64, error) {
 	}
 	defer destination.Close()
 	return io.Copy(destination, source)
-}
-
-// CleanDir removes all regular files under dir
-func CleanDir(dir string) error {
-	d, err := os.Open(dir)
-	if err != nil {
-		return err
-	}
-	defer d.Close()
-	names, err := d.Readdirnames(-1)
-	if err != nil {
-		return err
-	}
-	if len(names) == 0 {
-		fmt.Printf("dir \"%s\" is empty\n", dir)
-	}
-	var success []string
-	var failure []string
-	var errorSlice []error
-	for _, name := range names {
-		path := filepath.Join(dir, name)
-		// Validate path
-		fileInfo, err := os.Stat(path)
-		if err != nil {
-			failure = append(failure, path)
-			errorSlice = append(errorSlice, err)
-			continue
-		}
-		if !fileInfo.Mode().IsRegular() {
-			failure = append(failure, path)
-			errorSlice = append(errorSlice, errors.New("not a regular file"))
-			continue
-		}
-
-		err = os.RemoveAll(path)
-		if err != nil {
-			failure = append(failure, path)
-			errorSlice = append(errorSlice, err)
-			continue
-		}
-
-		success = append(success, path)
-	}
-
-	// write out
-	if len(success) != 0 {
-		fmt.Println("Successfully deleted following certificates from Trust Store:")
-		for _, p := range success {
-			fmt.Println(p)
-		}
-	}
-	if len(failure) != 0 {
-		fmt.Println("Failed to delete following certificates from Trust Store:")
-
-		for ind := range failure {
-			fmt.Printf("%s, with error \"%s\"\n", failure[ind], errorSlice[ind])
-		}
-	}
-
-	return nil
 }
